Fix nested zod types in the field chain
