+++ .changeset/lovely-clouds-joke.md
@@ -1,0 +1,5 @@
+---
+"@stevent-team/react-zoom-form": patch
+---
+
+Remove lib export of internal register function
