+++ .changeset/popular-pianos-wash.md
@@ -1,0 +1,5 @@
+---
+"@stevent-team/react-zoom-form": patch
+---
+
+Properly pass through the correct Zod schema types to the field chain using a second InnerSchema type param
